// src/types/castomHooks.types.ts
export interface INewData {
	dayWeek: string;
	dayNum: string;
	month: string;
	year: string;
	time: string;
}

export interface IRusDate {
	week: string;
	month: string;
}

export interface IObjectWeek {
	Mon: string;
	Tue: string;
	Wed: string;
	Thu: string;
	Fri: string;
	Sat: string;
	Sun: string;
}

export interface IObjectMonth {
	Jan: string;
	Feb: string;
	Mar: string;
	Apr: string;
	May: string;
	Jun: string;
	Jul: string;
	Aug: string;
	Sep: string;
	Oct: string;
	Nov: string;
	Dec: string;
}